@@ -17,31 +17,32 @@
     "url": "https://github.com/GatorIO/gator-web.git"
   },
   "engines": {
-    "node": ">=4.2.0"
+    "node": ">=8.7.0"
   },
   "dependencies": {
-    "@types/node": "^6.0.68",
-    "@types/async": "^2.0.40",
-    "async": "^2.3.0",
+    "@types/async": "^2.4.1",
+    "@types/connect-flash": "^0.0.34",
+    "@types/ejs": "^2.6.3",
+    "@types/express-session": "^1.15.12",
+    "@types/node": "^10.14.5",
+    "@types/restify": "^4.3.6",
+    "async": "^2.6.2",
     "body-parser": "*",
-    "@types/connect-redis": "^0.0.5",
-    "connect-redis": "^3.2.0",
-    "cookie-parser": "*",
-    "@types/ejs": "^2.3.33",
-    "ejs": "^2.5.5",
-    "@types/connect-flash": "^0.0.31",
     "connect-flash": "^0.1.1",
-    "express": "^4.14.0",
-    "@types/express-session": "^0.0.32",
-    "express-session": "^1.15.2",
-    "json-bigint": "^0.2.1",
-    "@types/lru-cache": "^2.5.32",
-    "lru-cache": "^4.0.2",
-    "gator-api": "GatorIO/gator-api",
-    "gator-utils": "GatorIO/gator-utils",
-    "@types/restify": "^2.0.42",
-    "restify": "^4.3.0",
-    "serve-favicon": "*"
+    "cookie-parser": "*",
+    "ejs": "^2.6.1",
+    "express": "^4.16.4",
+    "express-session": "^1.16.1",
+    "json-bigint": "^0.3.0",
+    "gator-api": "GatorIO/gator-api#2.x",
+    "gator-score": "gator-score",
+    "gator-utils": "GatorIO/gator-utils#2.x",
+    "restify": "^4.3.4",
+    "serve-favicon": "*",
+    "typescript": "^3.5.2"
   },
-  "license" : "MIT"
+  "license" : "MIT",
+  "scripts": {
+    "install": "tsc"
+  }
 }
