@@ -1,11 +1,8 @@
 import utils = require("gator-utils");
-import Utils = require("../lib/utils");
-import web = require("gator-web");
 import express = require('express');
 import restify = require('restify');
 import api = require('gator-api');
-import dictionaries = require('../lib/dictionaries');
-import {IApplication} from "gator-web";
+import {IApplication} from "../lib/index";
 
 export function getStations(req, res, application: IApplication) {
 
