@@ -2,7 +2,7 @@ import utils = require("gator-utils");
 import express = require('express');
 import restify = require('restify');
 import api = require('gator-api');
-import {IApplication} from "gator-web";
+import {IApplication} from "../lib";
 
 /*
  Set up routes - this script handles functions required for managing the API
@@ -10,13 +10,81 @@ import {IApplication} from "gator-web";
 
 export function setup(app: express.Application, application: IApplication, callback) {
 
+    //  proxy client api calls to api-host
+    app.post('/apiproxy', application.enforceSecure, api.authenticateNoRedirect, function (req: express.Request, res: express.Response) {
+
+        if (!req.body || !req.body.verb || !req.body.path) {
+            api.REST.sendError(res, new api.errors.MissingParameterError());
+        } else {
+
+            let client = api.sessionClient(req);
+
+            switch (req.body.verb.toUpperCase()) {
+
+                case 'DEL':
+                    client.del(req.body.path, function(err, apiRequest: restify.Request, apiResponse: restify.Response) {
+                        api.REST.sendConditional(res, err);
+                    });
+                    break;
+
+                case 'GET':
+                    client.get(req.body.path, function(err, apiRequest: restify.Request, apiResponse: restify.Response, result: any) {
+                        api.REST.sendConditional(res, err, result);
+                    });
+                    break;
+
+                case 'POST':
+                    client.post(req.body.path, req.body.data, function(err, apiRequest: restify.Request, apiResponse: restify.Response, result: any) {
+                        api.REST.sendConditional(res, err, result);
+                    });
+                    break;
+
+                case 'PATCH':
+                    client.patch(req.body.path, req.body.data, function(err, apiRequest: restify.Request, apiResponse: restify.Response, result: any) {
+                        api.REST.sendConditional(res, err, result);
+                    });
+                    break;
+
+                case 'PUT':
+                    client.put(req.body.path, req.body.data, function(err, apiRequest: restify.Request, apiResponse: restify.Response, result: any) {
+                        api.REST.sendConditional(res, err, result);
+                    });
+                    break;
+
+                default:
+                    api.REST.sendError(res, new api.errors.BadRequestError('No such verb'));
+            }
+        }
+    });
+
     app.get('/developer/rest/:id', application.enforceSecure, function (req: express.Request, res: express.Response) {
 
-        res.render('./developer/swagger', {
-            req: req,
-            application: application,
-            dev: utils.config.dev(),
-            spec: utils.config.settings().apiUrl + '/' + utils.config.settings().apiVersion + '/' + req.params['id'] + '.js'
+        let specUrl = utils.config.settings().apiUrl + '/' + utils.config.settings().apiVersion + '/' + req.params['id'] + '.json';
+
+        utils.getUrlText(specUrl, (err, spec) => {
+
+            if (err)
+                spec = {};
+            else {
+
+                try {
+                    spec = JSON.parse(spec);
+                } catch(err) {
+                    spec = {};
+                }
+            }
+
+            if (req.query.env == 'local') {
+                spec.host = '127.0.0.5:8080';
+                spec.schemes = [ 'http' ];
+            }
+
+            res.render('./developer/swagger', {
+                req: req,
+                application: application,
+                dev: utils.config.dev(),
+                spec: spec
+            });
         });
     });
 
@@ -24,8 +92,7 @@ export function setup(app: express.Application, application: IApplication, callb
 
         //  check for a remote login
         if (req.query.accessToken) {
-            api.setSessionCookie(res, req.query.accessToken);
-            res.redirect(application.branding.postLoginUrl);
+            res.redirect(utils.appendQueryString(application.branding.postLoginUrl, 'accessToken', req.query.accessToken));
             return;
         }
 
@@ -42,7 +109,7 @@ export function setup(app: express.Application, application: IApplication, callb
         api.login(req.body['username'], req.body['password'], application.settings.appId, function(err, authObject) {
 
             if (!err)
-                req.session.auth = authObject;
+                api.setSessionAuth(req, authObject);
 
             api.REST.sendConditional(res, err, null, 'success');
         });
@@ -94,8 +161,7 @@ export function setup(app: express.Application, application: IApplication, callb
             if (err) {
                 api.REST.sendError(res, err);
             } else {
-
-                api.setSessionCookie(res, authObject.accessToken);
+                api.setSessionAuth(req, authObject);
                 api.REST.sendConditional(res, err);
             }
         });
@@ -103,7 +169,7 @@ export function setup(app: express.Application, application: IApplication, callb
 
     //  handle logout
     app.get('/logout', application.enforceSecure, function(req: any, res) {
-        api.logout(res);
+        api.logout(req, res);
     });
 
     callback();
