@@ -2,8 +2,8 @@ import utils = require("gator-utils");
 import express = require('express');
 import restify = require('restify');
 import api = require('gator-api');
-import {IApplication} from "gator-web";
 import lib = require('../lib/index');
+import {IApplication} from "../lib";
 
 /*
  Set up routes - this script handles functions required for managing dashboards
