@@ -2,7 +2,7 @@ import utils = require("gator-utils");
 import express = require('express');
 import restify = require('restify');
 import api = require('gator-api');
-import {IApplication} from "gator-web";
+import {IApplication} from "../lib";
 
 /*
  Set up routes - this script handles functions required for managing payments
